Amend aggregate skeleton to take care of uneven aggregate
When calling a function on aggregates, it may happen that those
aggregates have different numbers of sub-parts. In this case, we
can't use the fast-track slicer calls and need to go back to a
recursive approach. By going recursive, we'll end up on a mixed
native/aggregate call that will be handled by architecture specific code.

This patch amend aggregate by adding replication computing
meta-fonctions that help us decide which strategy to use on
arbitrary aggregate calls.