Provide an extension system to decorators
Decorators may need to be combined so we needed a way to
make decorator type open-ended and easily specified.

This is done by having all decorator be an instance of a template
type decorated<T(A...)> that allow for decorators to aggregate info
when combined so that function specializations can inspect those
combos and act upon them.

Cleanup and small scale adaptations of existing code has been performed.