Cleanup PCH supports
Recent readings on PCH usage via CMake lead me to believe we were doing a few unnecessary steps. This is a cleaner version of PCH support. If it should normally not affect build time (the old version was working just awkwardly), this makes the PCH handling easier to extend later on.